pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "prompt-optimizer-env"
version = "1.0.0"
description = "Prompt Optimization RL Environment — Scaler OpenEV Hackathon"
requires-python = ">=3.11"
dependencies = [
    "openenv-core",
    "httpx>=0.27.0",
    "pydantic>=2.0",
    "uvicorn[standard]>=0.29.0",
    "fastapi>=0.111.0",
    "openai>=1.30.0",
]

[project.optional-dependencies]
dev = ["pytest", "pytest-asyncio", "httpx"]

[tool.hatch.build.targets.wheel]
packages = ["."]
